EKIR-573 Move Accessibility to RWPM
