+++ api/opds/opds2.py
@@ -18,7 +18,6 @@ from pydantic import (
 from typing_extensions import Self
 
 from api.opds import rwpm
-from api.opds.accessibility import AccessibilityDataExtension
 from api.opds.base import BaseOpdsModel
 from api.opds.palace import PalacePublicationMetadata
 from api.opds.types.currency import CurrencyCode
@@ -204,9 +203,7 @@ class FeedMetadata(BaseOpdsModel):
     number_of_items: NonNegativeInt | None = Field(None, alias="numberOfItems")
 
 
-class PublicationMetadata(
-    PalacePublicationMetadata, rwpm.Metadata, AccessibilityDataExtension
-):
+class PublicationMetadata(PalacePublicationMetadata, rwpm.Metadata):
     """
     OPDS2 publication metadata.
